解决兼容问题，内存表不能存放TEXT类型，导致post_keyword表与post_keyword_mem表字段不一致，post_keyword_mem缺少部分字段，查询时只能明确查询post_id字段，但是post_id字段是Long类型，在数据库中是BIGINT(20)类型，直接查询post_id字段时，JPA将它解析为BigInteger类型，与Long类型不兼容，所以将类PostKeyword中的postId修改为BigInteger类型，并修改所有使用到的地方

+++ src/main/java/com/bupt/hiservice/entity/post/PostKeyword.java
@@ -1,5 +1,7 @@
 package com.bupt.hiservice.entity.post;
 
+import java.math.BigInteger;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Table;
@@ -20,7 +22,7 @@ public class PostKeyword extends BaseEntity {
 	private String keyword;
 
 	@Column(name = "post_id")
-	private Long postId;
+	private BigInteger postId;
 
 	public String getKeyword() {
 		return keyword;
@@ -30,11 +32,11 @@ public class PostKeyword extends BaseEntity {
 		this.keyword = keyword;
 	}
 
-	public Long getPostId() {
+	public BigInteger getPostId() {
 		return postId;
 	}
 
-	public void setPostId(Long postId) {
+	public void setPostId(BigInteger postId) {
 		this.postId = postId;
 	}
